+++ climpred/prediction.py
@@ -1,5 +1,4 @@
 import inspect
-import warnings
 
 import xarray as xr
 
@@ -13,6 +12,7 @@ from .comparisons import (
     PROBABILISTIC_PM_COMPARISONS,
 )
 from .constants import CLIMPRED_DIMS, CONCAT_KWARGS, M2M_MEMBER_DIM, PM_CALENDAR_STR
+from .exceptions import DimensionError
 from .logging import log_compute_hindcast_header, log_compute_hindcast_inits_and_verifs
 from .metrics import HINDCAST_METRICS, METRIC_ALIASES, PM_METRICS
 from .reference import historical, persistence
@@ -86,10 +86,24 @@ def _apply_metric_at_given_lead(
     return result
 
 
-def _get_metric_comparison_dim(metric, comparison, dim, kind):
+def _sanitize_str_to_list(dim):
+    """Make dim to list if string, pass if None else raise ValueError."""
+    if isinstance(dim, str):
+        dim = [dim]
+    elif isinstance(dim, list) or dim is None:
+        dim = dim
+    else:
+        raise ValueError(
+            f'Expected `dim` as `str`, `list` or None, found {dim} as type {type(dim)}.'
+        )
+    return dim
+
+
+def _get_metric_comparison_dim(initialized, metric, comparison, dim, kind):
     """Returns `metric`, `comparison` and `dim` for compute functions.
 
     Args:
+        initialized (xr.object): initialized dataset: init_pm or hind
         metric (str): metric or alias string
         comparison (str): Description of parameter `comparison`.
         dim (list of str or str): dimension to apply metric to.
@@ -100,16 +114,20 @@ def _get_metric_comparison_dim(metric, comparison, dim, kind):
         comparison (Comparison): comparison class.
         dim (list of str or str): corrected dimension to apply metric to.
     """
+    dim = _sanitize_str_to_list(dim)
+
     # check kind allowed
     is_in_list(kind, ['hindcast', 'PM'], 'kind')
-    # set default dim
+    # set default dim # TODO: should we still do this?
     if dim is None:
-        dim = 'init' if kind == 'hindcast' else ['init', 'member']
-    # check allowed dims
-    if kind == 'hindcast':
-        is_in_list(dim, ['member', 'init'], 'dim')
-    elif kind == 'PM':
-        is_in_list(dim, ['member', 'init', ['init', 'member']], 'dim')
+        dim = ['init'] if kind == 'hindcast' else ['init', 'member']
+
+    # check that initialized contains all dims from dim
+    if not set(dim).issubset(initialized.dims):
+        raise DimensionError(
+            f'`dim`={dim} is expected to be a subset of '
+            f'`initialized.dims`={initialized.dims}.'
+        )
 
     # get metric and comparison strings incorporating alias
     metric = METRIC_ALIASES.get(metric, metric)
@@ -127,34 +145,21 @@ def _get_metric_comparison_dim(metric, comparison, dim, kind):
         else PROBABILISTIC_PM_COMPARISONS
     )
     if metric.probabilistic:
-        if not comparison.probabilistic:
+        if not comparison.probabilistic and 'member' in initialized.dims:
             raise ValueError(
                 f'Probabilistic metric `{metric.name}` requires comparison '
                 f'accepting multiple members e.g. `{PROBABILISTIC_COMPARISONS}`, '
                 f'found `{comparison.name}`.'
             )
-        if dim != 'member':
-            warnings.warn(
+        if 'member' not in dim and 'member' in initialized.dims:
+            raise ValueError(
                 f'Probabilistic metric {metric.name} requires to be '
-                f'computed over dimension `dim="member"`. '
-                f'Set automatically.'
+                f'computed over dimension `member`, which is not found in {dim}.'
             )
-            dim = 'member'
     else:  # determinstic metric
         if kind == 'hindcast':
-            if dim == 'init':
-                # for thinking in real time # compute_hindcast renames init to time
-                dim = 'time'
-        elif kind == 'PM':
-            # prevent comparison e2c and member in dim
-            if (comparison.name == 'e2c') and (
-                set(dim) == set(['init', 'member']) or dim == 'member'
-            ):
-                warnings.warn(
-                    f'comparison `{comparison.name}` does not work on `member` in dims,'
-                    f' found {dim}, automatically changed to dim=`init`.'
-                )
-                dim = 'init'
+            # for thinking in real time as compute_hindcast renames init to time
+            dim = ['time' if d == 'init' else d for d in dim]
     return metric, comparison, dim
 
 
@@ -164,7 +169,7 @@ def compute_perfect_model(
     control,
     metric='pearson_r',
     comparison='m2e',
-    dim=None,
+    dim=['member', 'init'],
     add_attrs=True,
     **metric_kwargs,
 ):
@@ -197,7 +202,7 @@ def compute_perfect_model(
 
     # check args compatible with each other
     metric, comparison, dim = _get_metric_comparison_dim(
-        metric, comparison, dim, kind='PM'
+        init_pm, metric, comparison, dim, kind='PM'
     )
 
     forecast, verif = comparison.function(init_pm, metric=metric)
@@ -205,11 +210,10 @@ def compute_perfect_model(
     # in case you want to compute deterministic skill over member dim
     if (forecast.dims != verif.dims) and not metric.probabilistic:
         forecast, verif = xr.broadcast(forecast, verif)
-
     skill = metric.function(
         forecast, verif, dim=dim, comparison=comparison, **metric_kwargs
     )
-    if comparison.name == 'm2m':
+    if comparison.name == 'm2m' and M2M_MEMBER_DIM in skill.dims:
         skill = skill.mean(M2M_MEMBER_DIM)
     # Attach climpred compute information to skill
     if add_attrs:
@@ -274,7 +278,7 @@ def compute_hindcast(
             Verification metric over ``lead`` reduced by dimension(s) ``dim``.
     """
     metric, comparison, dim = _get_metric_comparison_dim(
-        metric, comparison, dim, kind='hindcast'
+        hind, metric, comparison, dim, kind='hindcast'
     )
     hind = convert_time_index(hind, 'init', 'hind[init]')
     verif = convert_time_index(verif, 'time', 'verif[time]')
@@ -308,7 +312,7 @@ def compute_hindcast(
     result = xr.concat(metric_over_leads, dim='lead', **CONCAT_KWARGS)
     result['lead'] = forecast['lead']
     # rename back to 'init'
-    if 'time' in result.dims:  # If dim is 'member'
+    if 'time' in result.dims:
         result = result.rename({'time': 'init'})
     # These computations sometimes drop coordinates along the way. This appends them
     # back onto the results of the metric.