+++ climpred/tests/test_metrics_perfect.py
@@ -6,6 +6,8 @@ from climpred.comparisons import HINDCAST_COMPARISONS, PM_COMPARISONS
 from climpred.metrics import HINDCAST_METRICS, METRIC_ALIASES, PM_METRICS
 from climpred.utils import get_metric_class
 
+xr.set_options(display_style='text')
+
 
 @pytest.mark.parametrize('how', ['constant', 'increasing_by_lead'])
 @pytest.mark.parametrize('comparison', PM_COMPARISONS)
@@ -37,13 +39,16 @@ def test_PerfectModelEnsemble_constant_forecasts(
     elif metric == 'threshold_brier_score':
         metric_kwargs = {'threshold': 0.5}
     else:
-        metric_kwargs = {'useless_kwargs': 'to_ignore'}
+        metric_kwargs = {}
     if Metric.probabilistic:
-        skill = pe.verify(metric=metric, comparison='m2c', **metric_kwargs)
+        dim = ['init', 'member']
+        skill = pe.verify(metric=metric, comparison='m2c', dim=dim, **metric_kwargs)
     else:
-        skill = pe.verify(metric=metric, comparison=comparison, **metric_kwargs)
-    perfect_skill = Metric.perfect
-    assert skill == perfect_skill
+        dim = 'init' if comparison == 'e2c' else ['init', 'member']
+        skill = pe.verify(
+            metric=metric, comparison=comparison, dim=dim, **metric_kwargs
+        )
+    assert skill == Metric.perfect
 
 
 @pytest.mark.parametrize('alignment', ['same_inits', 'same_verif', 'maximize'])
@@ -94,14 +99,22 @@ def test_HindcastEnsemble_constant_forecasts(
     elif metric == 'threshold_brier_score':
         metric_kwargs = {'threshold': 0.5}
     else:
-        metric_kwargs = {'useless_kwargs': 'to_ignore'}
+        metric_kwargs = {}
     if Metric.probabilistic:
         skill = he.verify(
-            metric=metric, comparison='m2o', alignment=alignment, **metric_kwargs
+            metric=metric,
+            comparison='m2o',
+            dim='member',
+            alignment=alignment,
+            **metric_kwargs
         )
     else:
+        dim = 'member' if comparison == 'm2o' else 'init'
         skill = he.verify(
-            metric=metric, comparison=comparison, alignment=alignment, **metric_kwargs
+            metric=metric,
+            comparison=comparison,
+            dim=dim,
+            alignment=alignment,
+            **metric_kwargs
         )
-    perfect_skill = Metric.perfect
-    assert skill == perfect_skill
+    assert skill == Metric.perfect